added VFS root path and memory caching configuration

+++ d912pxy/d912pxy_device.h
@@ -177,7 +177,7 @@ public:
 
 	void CopyOriginalDX9Data(IDirect3DDevice9* dev, D3DDEVICE_CREATION_PARAMETERS* origPars, D3DPRESENT_PARAMETERS* origPP);
 	void InitVFS();
-	void InitVFSitem(UINT id, const char* name, UINT memCache);
+	void InitVFSitem(UINT id, const char* name, UINT64 memCache);
 
 	void InitClassFields();
 	void InitThreadSyncObjects();
